requirements.txt
asgiref==3.4.1
Django==4.0
djangorestframework==3.12.4
gunicorn==20.1.0
sqlparse==0.4.2
whitenoise==5.3.0
